Slurm scripts updates for mcptrue

+++ data/mcp.py
@@ -167,7 +167,7 @@ class MCP_Generator_True(Base_Generator):
         clique_solutions = solver.solutions[0]
         l_b = [adjacency_matrix_to_tensor_representation(W) for W in l_adjacency]
         l_k = [MCP_Generator.mcp_ind_to_adj(elt, self.n_vertices) for elt in clique_solutions]
-        l_data = [(B,K) for (B,K) in zip(l_b,l_k)]
+        l_data = [(B,K.to(int)) for (B,K) in zip(l_b,l_k)]
         return l_data
 
     def compute_example(self):
